# runtime: optimize time_now performance for linux amd64