--- a/src/runtime/timeasm.go
+++ b/src/runtime/timeasm.go
@@ -4,8 +4,8 @@
 
 // Declarations for operating systems implementing time.now directly in assembly.
 
-//go:build windows
-// +build windows
+//go:build windows || (linux && amd64)
+// +build windows linux,amd64
 
 package runtime