Currently time_now is implemented with two calls: `walltime` and `nanotime`.
The two calls both need to switch to g0 stack for vdso call and switch back.
So merge the two calls into one can avoid the extra cost once.
Benchmark shows a ~17% gain in performance.

```
name   old time/op  new time/op  delta
Now-8  40.9ns ± 9%  33.8ns ± 4%  -17.26%  (p=0.000 n=10+9)
```